pyproject.toml
[project]
name = "lol-esports-warehouse"
version = "0.1.0"
description = "Fetch professional esports game data from the Riot Games API"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
]
